refactor: move MonocolorLed into its own file
