+++ crates/sbd-gen-schema/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod ariel;
 pub mod common;
+pub mod led;
 pub mod riot;
 
 use std::collections::BTreeSet;
@@ -13,6 +14,8 @@ use crate::{
     common::StringOrVecString,
     riot::{Riot, RiotTargetExt},
 };
+
+pub use led::MonocolorLed;
 
 const fn default_version() -> Version {
     semver::Version::new(0, 2, 0)
@@ -94,16 +97,6 @@ impl Target {
     pub fn has_host_facing_uart(&self) -> bool {
         self.uarts.iter().any(|u| u.host_facing)
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct MonocolorLed {
-    pub pin: String,
-    pub color: Option<String>,
-    pub active: Option<PinActive>,
-    #[serde(default)]
-    pub aliases: Vec<String>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
